Respect link option in fallback file renderer (fix #2016)
