+++ application/src/Media/FileRenderer/FallbackRenderer.php
@@ -1,14 +1,6 @@
 <?php
 namespace Omeka\Media\FileRenderer;
 
-use Omeka\Api\Representation\MediaRepresentation;
-use Laminas\View\Renderer\PhpRenderer;
-
-class FallbackRenderer implements RendererInterface
+class FallbackRenderer extends AbstractRenderer
 {
-    public function render(PhpRenderer $view, MediaRepresentation $media,
-        array $options = []
-    ) {
-        return $view->hyperlink($media->filename(), $media->originalUrl());
-    }
 }
