+++ application/src/Media/FileRenderer/ThumbnailRenderer.php
@@ -4,7 +4,7 @@ namespace Omeka\Media\FileRenderer;
 use Omeka\Api\Representation\MediaRepresentation;
 use Laminas\View\Renderer\PhpRenderer;
 
-class ThumbnailRenderer implements RendererInterface
+class ThumbnailRenderer extends AbstractRenderer
 {
     public function render(PhpRenderer $view, MediaRepresentation $media,
         array $options = []
@@ -25,27 +25,5 @@ class ThumbnailRenderer implements RendererInterface
         $title = $media->displayTitle();
 
         return sprintf('<a href="%s" title="%s">%s</a>', $view->escapeHtml($url), $view->escapeHtml($title), $img);
-    }
-
-    /**
-     * Get the URL for the given type of link
-     *
-     * @param MediaRepresentation $media
-     * @param string $linkType Type of link: 'original', 'item', and 'media' are valid
-     * @throws \InvalidArgumentException On unrecognized $linkType
-     * @return string
-     */
-    protected function getLinkUrl(MediaRepresentation $media, $linkType)
-    {
-        switch ($linkType) {
-            case 'original':
-                return $media->originalUrl();
-            case 'item':
-                return $media->item()->url();
-            case 'media':
-                return $media->url();
-            default:
-                throw new \InvalidArgumentException(sprintf('Invalid link type "%s"', $linkType));
-        }
     }
 }
