+++ application/src/Media/FileRenderer/VideoRenderer.php
@@ -4,7 +4,7 @@ namespace Omeka\Media\FileRenderer;
 use Omeka\Api\Representation\MediaRepresentation;
 use Laminas\View\Renderer\PhpRenderer;
 
-class VideoRenderer implements RendererInterface
+class VideoRenderer extends AbstractRenderer
 {
     const DEFAULT_OPTIONS = [
         'controls' => true,
